CRMsoftware/requirements.txt
django
django-jalali
django-rest-framework